perform-tests on Tools class for the added methods
